requirements.txt
streamlit>=1.28.0
pandas>=1.5.3
python-docx>=0.8.11
PyPDF2>=3.0.0
langchain>=0.0.267
langchain_openai>=0.0.2
python-dotenv>=1.0.0